requirements.txt
appdirs==1.4.4
# black==21.5b1
cffi==1.14.5
click==8.0.0
colorama==0.4.4
cryptography==3.4.7
dnspython==1.16.0
Faker==8.1.3
Flask==2.0.0
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
mypy-extensions==0.4.3
pathspec==0.8.1
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.4
python-dateutil==2.8.1
regex==2021.4.4
six==1.16.0
text-unidecode==1.3
toml==0.10.2
Werkzeug==2.0.0